requirements/requirements.txt
django==4.0.4
djangorestframework==3.13.1
requests==2.28.0
djangorestframework-simplejwt==5.2.0
